backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
cffi==1.14.4
cryptography==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-lafv==0.0.2
ecdsa==0.14.1
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-jose==3.2.0
pytz==2020.5
PyYAML==5.3.1
rsa==4.7
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
